requirements.txt
fastapi
uvicorn
python-multipart
docling
pandas
openpyxl
google-genai
python-dotenv
pydantic
ollama
torch

